+++ pysui/abstracts/client_config.py
@@ -14,10 +14,19 @@
 """Client Configuration Abstraction."""
 from abc import ABC, abstractmethod
 from .client_types import AbstractType
+from .client_keypair import KeyPair
 
 
 class ClientConfiguration(ABC):
     """Base abstraction for managing a clients configuration."""
+
+    def __init__(self, keystore_file: str):
+        """Initialize base configuration properties."""
+        self._current_keystore_file = keystore_file
+        self._keypair_file: str = None
+        self._keypairs = {}
+        self._addresses = {}
+        self._address_keypair = {}
 
     @property
     @abstractmethod
@@ -28,3 +37,28 @@ class ClientConfiguration(ABC):
     @abstractmethod
     def active_address(self) -> AbstractType:
         """Return the active address from the client configuration."""
+
+    @property
+    def keystore_file(self) -> str:
+        """Get the kestore filename."""
+        return self._current_keystore_file
+
+    @property
+    def keystrings(self) -> list[str]:
+        """Get keypair strings managed by wallet."""
+        return list(self._keypairs)
+
+    @property
+    def addresses(self) -> list[str]:
+        """Get all the addresses."""
+        return list(self._addresses.keys())
+
+    def keypair_for_keystring(self, key_string: str) -> KeyPair:
+        """Get KeyPair for keystring."""
+        return self._keypairs[key_string]
+
+    def keypair_for_address(self, addy: AbstractType) -> KeyPair:
+        """Get the keypair for a given address."""
+        if addy.address in self._address_keypair:
+            return self._address_keypair[addy.address]
+        raise ValueError(f"{addy.address} is not known")