+++ pysui/sui/sui_rpc.py
@@ -14,15 +14,19 @@
 
 """Sui RPC Clients."""
 
+import json
+import base64
 from json import JSONDecodeError
-from typing import Any
+from typing import Any, Union
 import httpx
 from ..abstracts import SyncHttpRPC, RpcResult
+from .sui_types import SuiTxBytes
 from .sui_config import SuiConfig
-from .sui_builders import GetRpcAPI, SuiBaseBuilder
+from .sui_builders import GetRpcAPI, SuiBaseBuilder, ExecuteTransaction, SuiRequestType
 from .sui_apidesc import build_api_descriptors, SuiApi
 from .sui_txn_validator import validate_api
 from .sui_excepts import SuiRpcApiNotAvailable
+from .sui_tx_return_types import TxEffectResult
 
 
 class SuiRpcResult(RpcResult):
@@ -101,10 +105,8 @@ class SuiClient(SyncHttpRPC):
         """Perform argument validations."""
         return builder
 
-    def execute(self, builder: SuiBaseBuilder) -> Any:
+    def _execute(self, builder: SuiBaseBuilder) -> Union[SuiRpcResult, Exception]:
         """Execute the builder construct."""
-        if not builder.method in self._rpc_api:
-            raise SuiRpcApiNotAvailable(builder.method)
         parm_results = [y for x, y in validate_api(self._rpc_api[builder.method], builder)]
         jblock = self._generate_data_block(builder.data_dict, builder.method, parm_results)
         # import json
@@ -124,6 +126,40 @@ class SuiClient(SyncHttpRPC):
             return SuiRpcResult(False, f"JSON Decoder Error {jexc.msg}", vars(jexc))
         except httpx.ReadTimeout as hexc:
             return SuiRpcResult(False, "HTTP read timeout error", vars(hexc))
+
+    def execute(self, builder: SuiBaseBuilder) -> Union[SuiRpcResult, Exception]:
+        """Execute the builder construct."""
+        if not builder.method in self._rpc_api:
+            raise SuiRpcApiNotAvailable(builder.method)
+        if not builder.txn_required:
+            return self._execute(builder)
+        return self._signed_execution(builder)
+
+    def _signed_execution(self, builder: SuiBaseBuilder) -> Union[SuiRpcResult, Exception]:
+        """Subit base transaction, sign valid result and execute."""
+        result = self._execute(builder)
+        if result.is_ok():
+            result = result.result_data
+            if "error" in result:
+                return SuiRpcResult(False, result["error"]["message"], None)
+            kpair = self.config.keypair_for_address(builder.authority)
+            b64tx_bytes = result["result"]["txBytes"]
+            builder = ExecuteTransaction()
+            builder.set_pub_key(kpair.public_key).set_tx_bytes(SuiTxBytes(b64tx_bytes)).set_signature(
+                kpair.private_key.sign(base64.b64decode(b64tx_bytes))
+            ).set_sig_scheme(kpair.scheme).set_request_type(SuiRequestType.WAITFORLOCALEXECUTION)
+            # builder = DryRunTransaction()
+            # builder.set_pub_key(kpair.public_key).set_tx_bytes(SuiTxBytes(b64tx_bytes)).set_signature(
+            #     kpair.private_key.sign(base64.b64decode(b64tx_bytes))
+            # ).set_sig_scheme(kpair.scheme)
+            result = self.execute(builder)
+            if result.is_ok():
+                if "error" in result.result_data:
+                    return SuiRpcResult(False, result.result_data["error"]["message"], None)
+                # print(json.dumps(result.result_data["result"], indent=2))
+                tx_result = TxEffectResult.from_dict(result.result_data["result"])
+                return SuiRpcResult(tx_result.succeeded, tx_result.status, tx_result)
+        return result
 
     @property
     def rpc_api_names(self) -> list[str]: