+++ CHANGELOG.md
@@ -6,9 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-
 - CHANGELOG.md
+- Hydrate object results from Sui `sui_executeTransaction` API (see sui/sui_tx_return_types.py)
 
 ### Changed
 - Absolute to relative imports
-- Padding byte `0` to secp256k1 signatures to work with Sui
+- Fix for secp256k1 signing [issue](https://github.com/FrankC01/pysui/issues/12)
+- Moved signed transaction handling to sui/sui_rpc.py from faux_wallet
+- Moved configuration management to ClientConfig and SuiConfig from faux_wallet